+++ server/irc/irc-client-cache.js
@@ -37,67 +37,420 @@
 
   const path = require('path');
   const fs = require('fs');
+  const vars = require('./irc-client-vars');
 
   const credentials = JSON.parse(fs.readFileSync('./credentials.json', 'utf8'));
 
-  // size in lines of text
-  const cacheSize = 100;
-  let cacheInPointer = 0;
-  let cacheArray = [];
-  for (let i = 0; i < cacheSize; i++) {
-    cacheArray.push(null);
-  }
+  // ------------------------------------------------
+  // Example of data structure for message cache
+  //
+  // cachedArrays = {
+  //   default: [
+  //     'server message',
+  //     'server message',
+  //     'server message'
+  //   ]
+  //   '#mychannel': [
+  //     'channel message',
+  //     'channel message'
+  //   ]
+  // };
+  // cachedInPointers = {
+  //   default: 3,
+  //   '#mychannel': 2
+  // };
 
+  // size in lines of text for each message cache buffer
+  const cacheSize = 100; // messages
+  const maximumChannelCaches = 5; // IRC channel count (excluding default cache)
+
+  // Placeholder for primary cache variables
+  let cachedArrays = null;
+  let cachedInPointers = null;
+
+  /**
+   * Erase and re-initialize message cache buffers
+   * This is an external API method for this module
+   */
   const eraseCache = function () {
     // console.log('Cache Erased');
-    cacheInPointer = 0;
-    cacheArray = [];
+    cachedInPointers = {
+      default: 0
+    };
+    cachedArrays = {
+      default: []
+    };
     for (let i = 0; i < cacheSize; i++) {
-      cacheArray.push(null);
+      cachedArrays.default.push(null);
     }
   };
 
-  const addMessage = function (message) {
-    if (Buffer.isBuffer(message)) {
-      cacheArray[cacheInPointer] = Buffer.from(message);
-    } else if (typeof message === 'string') {
-      cacheArray[cacheInPointer] = Buffer.from(message, 'utf8');
-    }
-    cacheInPointer++;
-    if (cacheInPointer >= cacheSize) cacheInPointer = 0;
-  };
+  // On program start, initialize the message cache buffers
+  eraseCache();
 
-  const allMessages = function () {
-    const outArray = [];
-    let cacheOutPointer = cacheInPointer;
-    for (let i = 0; i < cacheSize; i++) {
-      if ((cacheArray[cacheOutPointer]) && (cacheArray[cacheOutPointer].length > 0)) {
-        // Option 1 to send array of utf8 encoded Buffer objects
-        outArray.push(cacheArray[cacheOutPointer]);
-        // Option 2 to send array of utf8 strings
-        // outArray.push(cacheArray[cacheOutPointer].toString('utf8'));
+  // ----------------------
+  // Notes on QUIT messages
+  // ----------------------
+  // In the IRC protocol, QUIT message do not contain an IRC channel names.
+  //
+  // Example: "nick!user@host QUIT :Reason for quitting"
+  //
+  // It is up to the IRC client to identify all channels where the QUIT nick is present.
+  // This is done by maintaining a list of IRC channels in ircState.channels[].
+  // Each IRC channel also maintains a list of members in ircState.channelStates[].names[]
+  // In a channel window, QUIT activities are display if the member was present in the list.
+  //
+  // In the case where the web browser reloads content from the cache,
+  // the member list may be out of date. Nicknames may or may not
+  // still be present in the channel. Therefore the channel name must be stored.
+  //
+  // A custom IRC message type "cachedQUIT" has been created. This is similar
+  // to a normal QUIT message except the channel name is added.
+  // The message is duplicated for each IRC channel.
+  //
+  // Example:  "nick!user@host cachedQUIT #channel :Reason for quitting"
+  //
+  // As real time messages are received from the IRC server, standard QUIT messages
+  // are parsed directly. When the browser content is deleted and restored from
+  // the IRC message cache, the alternate cachedQUIT messages are used.
+  //
+
+  /**
+   * Parse IRC message to determine if it is QUIT command.
+   * In the case of a QUIT command, generate an Array containing
+   * a list of IRC channels where the QUIT nickname has memebership.
+   *
+   * @param {Buffer} message - IRC server message encoded as UTF-8 Buffer.
+   * @returns {Object} - Returns object with isQuit flag and Array of IRC channels
+   */
+  function _parseQuitChannels (message) {
+    // First convert to a UTF-8 string
+    let messageUtf8 = null;
+    if (Buffer.isBuffer(message)) messageUtf8 = message.toString('utf8');
+    if (typeof message === 'string') messageUtf8 = message;
+    // then separate the string into an array of words
+    const messageWords = messageUtf8.split(' ');
+    if (messageWords.length > 1) {
+      // check if the command word is QUIT?
+      if (messageWords[2].toUpperCase() === 'QUIT') {
+        // Yes, it is a QUIT command, next get the associated IRC nickname
+        const nick = messageWords[1].split('!')[0].replace(':', '');
+        const channels = [];
+        if (vars.ircState.channelStates.length > 0) {
+          // Iterate through each IRC channel
+          for (let i = 0; i < vars.ircState.channelStates.length; i++) {
+            const channelStateObj = vars.ircState.channelStates[i];
+            if (channelStateObj.names.length > 0) {
+              // Iterate through each nickname in the channel names list
+              for (let j = 0; j < channelStateObj.names.length; j++) {
+                // if nickname is a member of the IRC channel, add it to the array
+                if (channelStateObj.names[j].toLowerCase() === nick.toLocaleLowerCase()) {
+                  channels.push(channelStateObj.name);
+                }
+              } // next j
+            } // length > 0
+          } // next i
+        } // length > 0
+        //
+        // return the array of membership channel names
+        return {
+          isQuit: true,
+          channels: channels
+        };
+      } else {
+        // Case of not QUIT message, should be displayed
+        return {
+          isQuit: false,
+          channels: []
+        };
       }
-      cacheOutPointer++;
-      if (cacheOutPointer >= cacheSize) cacheOutPointer = 0;
+    } else {
+      // case of un-parsable message, should be displayed
+      return {
+        isQuit: false,
+        channels: []
+      };
     }
-    return outArray;
-  };
+  } // _parseQuitChannels()
 
-  const cacheInfo = function () {
-    let size = 0;
-    let lines = 0;
-    if (cacheArray.length > 0) {
-      for (let i = 0; i < cacheArray.length; i++) {
-        if ((cacheArray[i]) && (cacheArray[i].length > 0)) {
-          lines++;
-          size += cacheArray[i].length;
+  //
+  // _channelCommands is a list of IRC server commands
+  // that should be segregated to a dedicated channel cache buffer.
+  //
+  // QUIT is not included due to special case handling
+  const _channelCommands = [
+    'JOIN',
+    'KICK',
+    'MODE',
+    'NOTICE',
+    'PART',
+    'PRIVMSG',
+    'TOPIC'
+  ];
+
+  /**
+   *
+   * @param {Buffer} message - IRC server message encoded as UTF-8 Buffer.
+   * @returns {Boolean} - Returns IRC channel name associated with command, or else returns null
+   */
+  function _extractChannel (message) {
+    let channel = null;
+    // First convert to a UTF-8 string
+    let messageUtf8 = null;
+    if (Buffer.isBuffer(message)) messageUtf8 = message.toString('utf8');
+    if (typeof message === 'string') messageUtf8 = message;
+    // then separate the string into an array of words
+    const messageWords = messageUtf8.split(' ');
+    if (messageWords.length > 2) {
+      if (_channelCommands.indexOf(messageWords[2].toUpperCase()) >= 0) {
+        // Check if the command word is JOIN?
+        // In the RFC protocol, the JOIN channel name starts with a colon :
+        // In the case of JOIN messages, remove the ":"
+        if (messageWords[2].toUpperCase() === 'JOIN') {
+          channel = messageWords[3].replace(':', '');
+        } else {
+          channel = messageWords[3];
+        }
+        //
+        // If this IRC client does not have membership in the channel specified,
+        // set the channel name to null to use the default buffer.
+        // Except, if this is a JOIN command for this client to join a new channel
+        // for the first time. In that case keep and return the channel name so the
+        // message cache buffer is created with JOIN as the first message in the cache.
+        //
+        // This is a negative (!( ... ))
+        //
+        if (!((messageWords[2].toUpperCase() === 'JOIN') &&
+          (vars.ircState.nickName === messageWords[1].split('!')[0].replace(':', '')))) {
+          if ((!channel) || (vars.ircState.channels.indexOf(channel) < 0)) {
+            channel = null;
+          }
         }
       }
     }
+    // return string with channel name or else return null
+    return channel;
+  }
+
+  const excludedMessageList = [
+    // JOIN channel Names list response
+    '353',
+    '366',
+    // /ADMIN
+    '256',
+    '257',
+    '258',
+    '259',
+    // /LIST response
+    '321',
+    '322',
+    '323',
+    // /WHO response
+    '315',
+    '352',
+    // /WHOIS response
+    '275',
+    '307',
+    '311',
+    '312',
+    '313',
+    '317',
+    '319',
+    '330',
+    '338',
+    '378',
+    '379',
+    '671'
+  ];
+
+  /**
+   * Function to extract the IRC command from the IRC message.
+   * If the command is included in the excluded command list
+   * return a flag to exclude this IRC command from storage into the cache
+   *
+   * @param {Buffer} message - IRC server message encoded as UTF-8 Buffer.
+   * @returns {Boolean} - Returns true if message NOT in the exclude list, else return false.
+   */
+  function _checkExclusionFilter (message) {
+    // First convert to a UTF-8 string
+    let messageUtf8 = null;
+    if (Buffer.isBuffer(message)) messageUtf8 = message.toString('utf8');
+    if (typeof message === 'string') messageUtf8 = message;
+    // then separate the string into an array of words
+    const messageWords = messageUtf8.split(' ');
+    if ((excludedMessageList.length > 0) && (messageWords.length > 1)) {
+      if (excludedMessageList.indexOf(messageWords[2].toUpperCase()) >= 0) {
+        // case of found, false === don't show message
+        return false;
+      } else {
+        return true; // true === show the message
+      }
+    } else {
+      // Cant check to go ahead
+      return true; // true === show the message
+    }
+  } // _checkExclusionFilter()
+
+  /**
+   * Internal function to add a single IRC message to the proper IRC channel cache buffer
+   *
+   * @param {String} message - String containing IRC channel name or 'default'.
+   * @param {Buffer} message - IRC server message encoded as UTF-8 Buffer.
+   */
+  function _addMessageToCacheBuffer (indexStr, messageAsBuf) {
+    if (indexStr !== 'default') {
+      // Check if a cache buffer exists, if not, create it and fill with null
+      if ((indexStr) &&
+        (Object.keys(cachedArrays).indexOf(indexStr) < 0) &&
+        // check if maximum number of channels is exceeded.
+        (Object.keys(cachedArrays).length < maximumChannelCaches + 1)) {
+        cachedInPointers[indexStr] = 0;
+        cachedArrays[indexStr] = [];
+        for (let i = 0; i < cacheSize; i++) {
+          cachedArrays[indexStr].push(null);
+        }
+      }
+    }
+    // Possibly maximum buffers exceeded, use default buffer
+    if (Object.keys(cachedArrays).indexOf(indexStr) < 0) indexStr = 'default';
+
+    cachedArrays[indexStr][cachedInPointers[indexStr]] = messageAsBuf;
+    cachedInPointers[indexStr]++;
+    if (cachedInPointers[indexStr] >= cacheSize) cachedInPointers[indexStr] = 0;
+  } // _addMessageToCacheBuffer()
+
+  /**
+   * This will add an IRC message to the IRC message cache.
+   * This is an external API method for this module
+   *
+   * @param {String} message - String containing IRC channel name or 'default'.
+   */
+  const addMessage = function (message) {
+    // Edge case, quit message does not contain channel name, may refer to multiple channels
+    // see parsedQuitChannels above...
+    //
+    // parsedQuitChannels contains an array of IRC channel names
+    // where the nickname has channel membership
+    const parsedQuitChannels = _parseQuitChannels(message);
+    // Is IRC message a QUIT message?
+    if (parsedQuitChannels.isQuit) {
+      // Special case of IRC QUIT message, one user can be in multiple channels at the same time
+      //
+      // iterate each array element (each cache buffer) and
+      // then duplicate the QUIT message to that cache buffer
+      parsedQuitChannels.channels.forEach(function (indexStr) {
+        // Modify the QUIT message before caching duplicate copies
+        // 1) Change "QUIT" to "cachedQUIT"
+        // 2) Insert channel name as string value
+        // Example:  "nick!user@host cachedQUIT #channel :Reason for quitting"
+        let messageAsStr = null;
+        if (Buffer.isBuffer(message)) messageAsStr = message.toString('utf8');
+        if (typeof message === 'string') messageAsStr = message;
+        const messageAsWords = messageAsStr.split(' ');
+        messageAsStr = '';
+        if (messageAsWords.length > 0) {
+          for (let i = 0; i < messageAsWords.length; i++) {
+            if (i === 0) {
+              messageAsStr += messageAsWords[i];
+            } else if (i === 2) {
+              messageAsStr += ' cachedQUIT ' + indexStr;
+            } else {
+              messageAsStr += ' ' + messageAsWords[i];
+            }
+          }
+        }
+        const messageAsBuf = Buffer.from(messageAsStr, 'utf8');
+        // save it to the cache buffer
+        if (Object.keys(cachedArrays).indexOf(indexStr) >= 0) {
+          // console.log('Adding QUIT to ' + indexStr + ' cache buffer');
+          _addMessageToCacheBuffer(indexStr, messageAsBuf);
+        } else {
+          // console.log('Adding QUIT to default cache buffer');
+          _addMessageToCacheBuffer('default', messageAsBuf);
+        }
+      });
+    } else {
+      // Case of not a QUIT message, this is a simple case.
+
+      // Type conversion to UTF-8 encoded Buffer
+      let messageAsBuf = null;
+      if (Buffer.isBuffer(message)) messageAsBuf = Buffer.from(message);
+      if (typeof message === 'string') messageAsBuf = Buffer.from(message, 'utf8');
+
+      const parsedChannel = _extractChannel(message);
+      if (_checkExclusionFilter(message)) {
+        if ((parsedChannel) && (parsedChannel.length > 0)) {
+          _addMessageToCacheBuffer(parsedChannel, messageAsBuf);
+        } else {
+          _addMessageToCacheBuffer('default', messageAsBuf);
+        }
+      }
+    } // isQuit
+  }; // addMessage()
+
+  // Debug: show cache buffers each 15 seconds.
+
+  // setInterval(() => {
+  //   Object.keys(cachedArrays).forEach(function (indexStr) {
+  //     cachedArrays[indexStr].forEach(function (line) {
+  //       if (line) {
+  //         console.log(indexStr + ': ' + line.toString('utf8'));
+  //       } else {
+  //         console.log(indexStr + ': null');
+  //       }
+  //     });
+  //   });
+  //   console.log(cachedInPointers);
+  // }, 15000);
+
+  /**
+   * Concatenate all cache buffers into an Array of IRC messages
+   * This is an external API method for this module
+   *
+   * @returns {Array} - Returns an Array of UTF-8 encoded Buffers
+   */
+  const allMessages = function () {
+    const outArray = [];
+    const arrayNameList = Object.keys(cachedArrays);
+    arrayNameList.forEach(function (indexStr) {
+      let cacheOutPointer = cachedInPointers[indexStr];
+      for (let i = 0; i < cacheSize; i++) {
+        if ((cachedArrays[indexStr][cacheOutPointer]) &&
+          (cachedArrays[indexStr][cacheOutPointer].length > 0)) {
+          // Option 1 to send array of utf8 encoded Buffer objects
+          outArray.push(cachedArrays[indexStr][cacheOutPointer]);
+          // Option 2 to send array of utf8 strings
+          // outArray.push(cachedArrays[indexStr][cacheOutPointer].toString('utf8'));
+        }
+        cacheOutPointer++;
+        if (cacheOutPointer >= cacheSize) cacheOutPointer = 0;
+      }
+    });
+    return outArray;
+  };
+
+  /**
+   * Returns data related to size of the message cach3
+   * This is an external API method for this module
+   *
+   * @returns {Object} - Returns Object with properties showing the size of the cache
+   */
+  const cacheInfo = function () {
+    let size = 0;
+    let lines = 0;
+    const arrayNameList = Object.keys(cachedArrays);
+    arrayNameList.forEach(function (indexStr) {
+      for (let i = 0; i < cacheSize; i++) {
+        if ((!(cachedArrays[indexStr][i] == null)) && (cachedArrays[indexStr][i].length > 0)) {
+          lines++;
+          size += cachedArrays[indexStr][i].length;
+        }
+      }
+    });
     const info = {};
     info.sizeCache = cacheSize;
-    info.sizeArray = cacheArray.length;
-    info.inPointer = cacheInPointer;
+    // info.sizeArray = cacheArray.length;
+    // info.inPointer = cacheInPointer;
     info.usedLines = lines;
     info.sizeInBytes = size;
     return info;
@@ -152,20 +505,25 @@
     // timestamp in seconds
     const timestamp = parseInt(now.getTime() / 1000);
 
-    const stringArray = [];
-    for (let i = 0; i < cacheSize; i++) {
-      if (cacheArray[i] == null) {
-        stringArray.push(null);
-      } else {
-        stringArray.push(cacheArray[i].toString('utf8'));
+    const stringArrays = {};
+    const arrayNameList = Object.keys(cachedArrays);
+    arrayNameList.forEach(function (indexStr) {
+      stringArrays[indexStr] = [];
+      for (let i = 0; i < cacheSize; i++) {
+        if (cachedArrays[indexStr][i] == null) {
+          stringArrays[indexStr].push(null);
+        } else {
+          stringArrays[indexStr].push(cachedArrays[indexStr][i].toString('utf8'));
+        }
       }
-    }
+    });
+
     const encodedCache = JSON.stringify(
       {
         timestamp: timestamp,
         hasBeenUsed: false,
-        cacheInPointer: cacheInPointer,
-        cacheArray: stringArray
+        cachedInPointers: cachedInPointers,
+        cachedArrays: stringArrays
       }
     );
     // Asynchronous File I/O
@@ -252,24 +610,35 @@
             }
             if ((decoded) && (!decoded.hasBeenUsed)) {
               if (timeNow - decoded.timestamp < cacheReloadWindowSeconds) {
-                if (
-                  ('cacheInPointer' in decoded) &&
-                  ('cacheArray' in decoded) &&
-                  (!(cacheInPointer == null)) &&
-                  (Array.isArray(decoded.cacheArray)) &&
-                  (decoded.cacheArray.length === cacheSize)) {
-                  cacheInPointer = decoded.cacheInPointer;
-                  const tempCacheArray = [];
-                  for (let i = 0; i < cacheSize; i++) {
-                    if (decoded.cacheArray[i] == null) {
-                      tempCacheArray.push(null);
-                    } else {
-                      // console.log(i, decoded.cacheArray[i]);
-                      tempCacheArray.push(Buffer.from(decoded.cacheArray[i], 'utf8'));
+                // Check for legacy file format, skip if found
+                // Prior to v0.1.14, 'cachedArrays' was 'cacheArray'
+                if (('cachedInPointers' in decoded) &&
+                  ('cachedArrays' in decoded)) {
+                  // Directly restore the circular buffer array pointers
+                  cachedInPointers = decoded.cachedInPointers;
+                  // List of array keys, example: ['default', '#channel1', '#otherchannel']
+                  const arrayNameList = Object.keys(cachedInPointers);
+                  // cycle through each array buffer
+                  arrayNameList.forEach(function (indexStr) {
+                    if ((!(cachedInPointers[indexStr] == null)) &&
+                      (Array.isArray(decoded.cachedArrays[indexStr])) &&
+                      (decoded.cachedArrays[indexStr].length === cacheSize)) {
+                      // Convert each UTF-8 string in the array to Buffer objects.
+                      cachedArrays[indexStr] = [];
+                      for (let i = 0; i < cacheSize; i++) {
+                        if (decoded.cachedArrays[indexStr][i] == null) {
+                          cachedArrays[indexStr].push(null);
+                        } else {
+                          // console.log(i, decoded.cacheArray[i]);
+                          cachedArrays[indexStr].push(
+                            Buffer.from(decoded.cachedArrays[indexStr][i], 'utf8'));
+                        }
+                      } // for (let i=...)
                     }
-                  }
-                  cacheArray = tempCacheArray;
+                  }); // forEach
                   console.log('Message cache loaded from file');
+                } else {
+                  console.log('Message cache: legacy format detected, not reloaded.');
                 }
               } else {
                 console.log('Message cache expired, not reloaded.');
@@ -278,10 +647,15 @@
                 {
                   timestamp: timeNow,
                   hasBeenUsed: true,
-                  cacheInPointer: null,
-                  cacheArray: null
+                  cachedInPointers: {
+                    default: 0
+                  },
+                  cachedArrays: {
+                    default: []
+                  }
                 }
               );
+              // Erase the old saved cache so it can not be used again.
               fs.writeFile(
                 cacheFilename,
                 encodedCache,
@@ -296,12 +670,15 @@
                     process.exit(1);
                   }
                 }
-              );
-            }
+              ); // fs.writeFile()
+            } // if ((decoded) && (!decoded.hasBeenUsed)) {
+          } else {
+            console.log(err || err.toString());
+            process.exit(1);
           }
-        }
-      );
-    })();
+        } // callback function
+      ); // fs.readfile
+    })(); // immediate function
   } // if (credentials.persistIrcMessageCache)
 
   module.exports = {